fix: code clean-up #3

+++ src/main/java/ru/practicum/shareit/booking/BookingServiceImpl.java
@@ -61,7 +61,7 @@ public class BookingServiceImpl implements BookingService {
     public BookingDto get(Long bookingId, Long userId) {
         Booking booking = bookingRepository.findById(bookingId).orElseThrow(() ->
                 new NoSuchObjectException("Booking not found"));
-        if (booking.getBooker().getId() == userId || booking.getOwner().getId() == userId) {
+        if (booking.getBooker().getId().longValue() == userId.longValue() || booking.getOwner().getId() == userId) {
             return BookingMapper.bookingToBookingDto(booking);
         } else {
             throw new NoSuchObjectException("Access denied.");
